src/requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==21.2.0  # https://github.com/benoitc/gunicorn
psycopg[c]==3.1.12  # https://github.com/psycopg/psycopg

# Django
# ------------------------------------------------------------------------------
django-anymail==10.1  # https://github.com/anymail/django-anymail

django-unfold==0.60.0
django-compressor==4.4 
django-debug-toolbar==4.2.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions

isoweek==1.3.3